AFF4_read now returns a ssize_t.

+++ tests/aff4_capi.cc
@@ -50,7 +50,7 @@ TEST_F(AFF4CAPI, Sample1URN) {
 
     char* buffer = (char*) malloc(33);
     memset(buffer, 0, 33);
-    int read = AFF4_read(handle, 0, buffer, 32, nullptr);
+    ssize_t read = AFF4_read(handle, 0, buffer, 32, nullptr);
     ASSERT_EQ(32, read);
     ASSERT_STREQ("\x33\xC0\x8E\xD0\xBC\x00\x7C\x8E\xC0\x8E\xD8\xBE\x00\x7C\xBF\x00\x06\xB9\x00\x02\xFC\xF3\xA4\x50\x68\x1C\x06\xCB\xFB\xB9\x04\x00", buffer);
 
@@ -72,7 +72,7 @@ TEST_F(AFF4CAPI, Sample2URN) {
     memset(buffer, 0, 33);
 
     // Start
-    int read = AFF4_read(handle, 0, buffer, 32, nullptr);
+    ssize_t read = AFF4_read(handle, 0, buffer, 32, nullptr);
     ASSERT_EQ(32, read);
     ASSERT_STREQ("\x33\xC0\x8E\xD0\xBC\x00\x7C\x8E\xC0\x8E\xD8\xBE\x00\x7C\xBF\x00\x06\xB9\x00\x02\xFC\xF3\xA4\x50\x68\x1C\x06\xCB\xFB\xB9\x04\x00", buffer);
 
@@ -100,7 +100,7 @@ TEST_F(AFF4CAPI, Sample3URN) {
     memset(buffer, 0, 33);
 
     // Start...
-    int read = AFF4_read(handle, 0, buffer, 32, nullptr);
+    ssize_t read = AFF4_read(handle, 0, buffer, 32, nullptr);
     ASSERT_EQ(32, read);
     ASSERT_STREQ("\x33\xC0\x8E\xD0\xBC\x00\x7C\x8E\xC0\x8E\xD8\xBE\x00\x7C\xBF\x00\x06\xB9\x00\x02\xFC\xF3\xA4\x50\x68\x1C\x06\xCB\xFB\xB9\x04\x00", buffer);
